Ajustando nome de classes
- Haviam classes com o nome duplicado
